chore(core): split cpp files to make it easier for compiler
